fix country area column type

+++ db/migrate/20150216184513_create_geonames_countries.rb
@@ -8,7 +8,7 @@ class CreateGeonamesCountries < ActiveRecord::Migration[5.2]
       t.string :fips
       t.string :country
       t.string :capital
-      t.integer :area # in sq. km
+      t.float :area # in sq. km
       t.integer :population
       t.string :continent
       t.string :tld
